feat: add WithdrawHistoryActiveTab component and integrate it into RevenueManagement
